requirements.txt
Django==5.0.6
django-heroku==0.3.1
gunicorn==22.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1
six==1.16.0
python-decouple==3.8
django-unfold==0.43.0
django-google-sso==6.2.1
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-decouple==3.8
opencv-python==4.10.0.84
firebase-admin==6.5.0
Pillow==10.3.0
requests==2.32.2
Faker==33.1.0
